objc type precision corrected

+++ Objective-C/Tests/ReplicatorTest+CustomConflict.m
@@ -755,7 +755,7 @@
     
     NSMutableArray* order = [NSMutableArray array];
     resolver = [[TestConflictResolver alloc] initWithResolver: ^CBLDocument* (CBLConflict* con) {
-        int count;
+        NSUInteger count;
         @synchronized (order) {
             [order addObject: con.localDocument.id];
             count = order.count;
